move FileWithType from utils to dtos - POM rest
