+++ invoice-tracker-backend/pom.xml
@@ -232,7 +232,6 @@
 						<exclude>net/talaatharb/**/*Application.*</exclude>
 						<exclude>net/talaatharb/**/*DTO.*</exclude>
 						<exclude>**/config/*</exclude>
-						<exclude>**/utils/*</exclude>
 					</excludes>
 				</configuration>
 				<executions>
